# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Assist Tix API Documentation",
        "title": "Assist Tix Documentation",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/organizers": {
            "get": {
                "description": "Get all organizer",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizer"
                ],
                "summary": "Get all organizer",
                "responses": {
                    "200": {
                        "description": "Organizers response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/lib.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.OrganizerResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create organizer",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizer"
                ],
                "summary": "Create organizer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name",
                        "name": "name",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Slug",
                        "name": "slug",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Logo",
                        "name": "logo",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organizer created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/lib.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    }
                }
            }
        },
        "/organizers/{organizerId}": {
            "get": {
                "description": "Get organizer by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizer"
                ],
                "summary": "Get organizer by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organizer ID",
                        "name": "organizerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Organizer response",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/lib.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.OrganizerResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update organizer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizer"
                ],
                "summary": "Update organizer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organizer ID",
                        "name": "organizerId",
                        "in": "path"
                    },
                    {
                        "description": "Create venue request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateOrganizerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Venue created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/lib.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete organizer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "organizer"
                ],
                "summary": "Delete organizer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Organizer ID",
                        "name": "organizerId",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Venue created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/lib.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/lib.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.OrganizerResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.UpdateOrganizerRequest": {
            "type": "object",
            "required": [
                "name",
                "slug"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "lib.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "default": true
                }
            }
        },
        "lib.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "default": false
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}